requirements.txt
fastapi==0.109.0
strawberry-graphql==0.217.0
uvicorn==0.27.0
pydantic==2.5.3
